@@ -30,12 +30,12 @@
   },
   "dependencies": {
     "@hmcts/info-provider": "^1.0.0",
-    "@hmcts/nodejs-healthcheck": "^1.7.0",
-    "@hmcts/nodejs-logging": "^3.0.2",
-    "@hmcts/properties-volume": "^0.0.13",
+    "@hmcts/nodejs-healthcheck": "^1.7.3",
+    "@hmcts/nodejs-logging": "^4.0.4",
+    "@hmcts/properties-volume": "^0.0.14",
     "@types/config": "^0.0.41",
     "@types/connect-redis": "^0.0.18",
-    "@types/cookie-parser": "^1.4.2",
+    "@types/cookie-parser": "^1.4.3",
     "@types/csurf": "^1.9.36",
     "@types/es6-promisify": "^6.0.1",
     "@types/express": "^4.17.13",
@@ -43,16 +43,17 @@
     "@types/glob": "^7.2.0",
     "@types/helmet": "^4.0.0",
     "@types/jest-when": "^3.5.0",
-    "@types/node": "^17.0.22",
+    "@types/json2csv": "^5.0.3",
     "@types/lodash": "^4.14.180",
+    "@types/node": "^17.0.22",
     "@types/nunjucks": "^3.2.1",
     "@types/require-directory": "^2.1.1",
     "@types/serve-favicon": "^2.5.1",
     "@types/session-file-store": "^1.2.2",
     "@types/uuid": "^8.3.4",
-    "applicationinsights": "^2.3.2",
+    "applicationinsights": "^2.3.3",
     "autobind-decorator": "^2.4.0",
-    "awilix": "^7.0.2",
+    "awilix": "^7.0.3",
     "axios": "^0.27.2",
     "config": "^3.3.7",
     "connect-redis": "^6.1.3",
@@ -65,9 +66,11 @@
     "govuk-frontend": "^4.0.1",
     "helmet": "^5.0.2",
     "jest-when": "^3.5.1",
+    "json2csv": "^5.0.7",
     "jwt-decode": "^3.1.2",
     "launchdarkly-node-server-sdk": "^6.4.1",
     "lodash": "^4.17.20",
+    "memfs": "^3.4.7",
     "npm-run-all": "^4.1.5",
     "nunjucks": "^3.2.2",
     "redis": "^3.1.2",
@@ -82,8 +85,7 @@
   },
   "devDependencies": {
     "@babel/core": "^7.17.7",
-    "@babel/preset-env": "^7.17.10",
-    "notifications-node-client": "^4.7.3",
+    "@babel/preset-env": "^7.18.2",
     "@codeceptjs/helper": "^1.0.2",
     "@types/copy-webpack-plugin": "^10.1.0",
     "@types/jest": "^27.4.1",
@@ -107,6 +109,7 @@
     "nock": "^13.2.4",
     "node-sass": "^7.0.1",
     "nodemon": "^2.0.16",
+    "notifications-node-client": "^4.7.3",
     "pa11y": "^6.1.1",
     "playwright": "^1.19.2",
     "sass-loader": "^12.6.0",
@@ -115,9 +118,9 @@
     "ts-jest": "^27.1.4",
     "ts-loader": "^9.3.0",
     "tslint": "6.1.3",
-    "webpack": "^5.70.0",
+    "webpack": "^5.73.0",
     "webpack-cli": "^4.9.2",
-    "webpack-dev-middleware": "^5.3.1",
+    "webpack-dev-middleware": "^5.3.3",
     "webpack-node-externals": "^3.0.0"
   },
   "resolutions": {
