@@ -3,7 +3,7 @@ import { Response } from 'express';
 import { RootController } from './RootController';
 import autobind from 'autobind-decorator';
 import asyncError from '../modules/error-handler/asyncErrorDecorator';
-import { EDIT_USER_URL, USER_DELETE_URL, USER_SUSPEND_URL } from '../utils/urls';
+import { EDIT_USER_URL, USER_SUSPEND_URL, USER_DELETE_URL, USER_DISABLE_SSO_URL } from '../utils/urls';
 
 @autobind
 export class UserActionsController extends RootController{
@@ -16,6 +16,8 @@ export class UserActionsController extends RootController{
       case 'suspend':
       case 'unsuspend':
         return res.redirect(307, USER_SUSPEND_URL);
+      case 'remove-sso':
+        return res.redirect(307, USER_DISABLE_SSO_URL);
       case 'delete':
         return res.redirect(307, USER_DELETE_URL);
     }
