@@ -10,11 +10,22 @@ import {
   USER_ACTIONS_URL,
   USER_DETAILS_URL,
   USER_SUSPEND_URL,
+  USER_DISABLE_SSO_URL,
+  ACCESSIBILITY_STATEMENT,
+  GENERATE_REPORT_URL,
+  DOWNLOAD_REPORT_URL,
   ADD_PRIVATE_BETA_SERVICE_URL,
-  ACCESSIBILITY_STATEMENT
 } from './utils/urls';
 import { FeatureFlags } from './app/feature-flags/FeatureFlags';
-import { BETA_ADD, BETA_DELETE, BETA_EDIT, BETA_SUSPEND, GAMMA_PRIVATE_BETA } from './app/feature-flags/flags';
+import {
+  BETA_ADD,
+  BETA_DELETE,
+  BETA_EDIT,
+  BETA_SUSPEND,
+  BETA_REMOVE_SSO,
+  GAMMA_PRIVATE_BETA,
+  GAMMA_GENERATE_REPORT
+} from './app/feature-flags/flags';
 
 export default function(app: Application): void {
   const featureFlags: FeatureFlags = app.locals.container.cradle.featureFlags;
@@ -31,5 +42,9 @@ export default function(app: Application): void {
   app.post(EDIT_USER_URL, featureFlags.toggleRoute(BETA_EDIT), app.locals.container.cradle.userEditController.post);
   app.post(USER_DELETE_URL, featureFlags.toggleRoute(BETA_DELETE), app.locals.container.cradle.userDeleteController.post);
   app.post(USER_SUSPEND_URL, featureFlags.toggleRoute(BETA_SUSPEND), app.locals.container.cradle.userSuspendController.post);
+  app.post(USER_DISABLE_SSO_URL, featureFlags.toggleRoute(BETA_REMOVE_SSO), app.locals.container.cradle.userSsoController.post);
   app.get(ACCESSIBILITY_STATEMENT, app.locals.container.cradle.accessibilityStatementController.get);
+  app.get(GENERATE_REPORT_URL, featureFlags.toggleRoute(GAMMA_GENERATE_REPORT), app.locals.container.cradle.generateReportController.get);
+  app.post(GENERATE_REPORT_URL, featureFlags.toggleRoute(GAMMA_GENERATE_REPORT), app.locals.container.cradle.generateReportController.post);
+  app.get(DOWNLOAD_REPORT_URL, featureFlags.toggleRoute(GAMMA_GENERATE_REPORT), app.locals.container.cradle.downloadReportController.get);
 }
