@@ -6,29 +6,30 @@ import {
 import {config as testConfig} from '../config';
 import * as Assert from 'assert';
 import {randomData} from './shared/random-data';
-import { BETA_EDIT, GAMMA_MFA } from '../../main/app/feature-flags/flags';
+import {BETA_EDIT, GAMMA_MFA} from '../../main/app/feature-flags/flags';
 
 Feature('Edit User');
 
 const PARENT_ROLE = randomData.getRandomRole();
 const PARENT_ROLE_WITH_MFA_ASSIGNABLE = randomData.getRandomRole();
-const ASSIGNABLE_CHILD_ROLE = randomData.getRandomRole();
+const ASSIGNABLE_CHILD_ROLE1 = randomData.getRandomRole();
+const ASSIGNABLE_CHILD_ROLE2 = randomData.getRandomRole();
 const INDEPENDANT_CHILD_ROLE = randomData.getRandomRole();
 const PARENT_ROLE_EMAIL = randomData.getRandomEmailAddress();
 
 const MFA_ENABLED_FLAG = 'enabled';
-const MFA_ENABLED_TEXT = 'Enabled';
-const MFA_DISABLED_TEXT = 'Disabled';
+const MFA_DISABLED_TEXT = 'DISABLED';
 const MFA_SECURITY_WARNING = 'Only disable MFA for a user if they have a \'justice.gov.uk\' or \'hmcts.net\' email address. Contact the information security team if you want to make an exception.';
 
 BeforeSuite(async () => {
   await createAssignableRoles(PARENT_ROLE);
   await createAssignableRoles(PARENT_ROLE_WITH_MFA_ASSIGNABLE);
-  await createAssignableRoles(ASSIGNABLE_CHILD_ROLE);
+  await createAssignableRoles(ASSIGNABLE_CHILD_ROLE1);
+  await createAssignableRoles(ASSIGNABLE_CHILD_ROLE2);
   await createAssignableRoles(INDEPENDANT_CHILD_ROLE);
   // Assigning self role with the child role so the this user can also delete same level users
-  await assignRolesToParentRole(PARENT_ROLE, [ASSIGNABLE_CHILD_ROLE, PARENT_ROLE]);
-  await assignRolesToParentRole(PARENT_ROLE_WITH_MFA_ASSIGNABLE, [ASSIGNABLE_CHILD_ROLE, PARENT_ROLE_WITH_MFA_ASSIGNABLE, testConfig.USER_ROLE_IDAM_MFA_DISABLED]);
+  await assignRolesToParentRole(PARENT_ROLE, [ASSIGNABLE_CHILD_ROLE1, PARENT_ROLE, ASSIGNABLE_CHILD_ROLE2]);
+  await assignRolesToParentRole(PARENT_ROLE_WITH_MFA_ASSIGNABLE, [ASSIGNABLE_CHILD_ROLE1, PARENT_ROLE_WITH_MFA_ASSIGNABLE, testConfig.USER_ROLE_IDAM_MFA_DISABLED]);
   await createUserWithRoles(PARENT_ROLE_EMAIL, testConfig.PASSWORD, testConfig.USER_FIRSTNAME, [testConfig.RBAC.access, PARENT_ROLE]);
 });
 
@@ -185,30 +186,26 @@ Scenario('I as a user should be able to edit roles only if I have the permission
     I.fillField('#search', activeUserEmail);
     I.click('Search');
     I.waitForText('User Details');
-
-    const assignedRoles = await I.grabTextFromAll('[id^=\'assigned-role\']');
-    Assert.equal(assignedRoles.includes(INDEPENDANT_CHILD_ROLE), true);
-    Assert.equal(assignedRoles.includes(ASSIGNABLE_CHILD_ROLE), false);
+    I.see(INDEPENDANT_CHILD_ROLE);
+    I.dontSee(ASSIGNABLE_CHILD_ROLE1);
 
     I.click('Edit user');
     I.waitForText('Edit User');
-    I.see(ASSIGNABLE_CHILD_ROLE);
+    I.see(ASSIGNABLE_CHILD_ROLE1);
     I.see(INDEPENDANT_CHILD_ROLE);
 
     //Checking the role which user doesn't have permission to assign is disabled
     const disabledRoles = await I.grabValueFromAll(locate('//input[@name=\'roles\' and @disabled]'));
     Assert.equal(disabledRoles.includes(INDEPENDANT_CHILD_ROLE), true);
 
-    I.click(ASSIGNABLE_CHILD_ROLE);
+    I.click(ASSIGNABLE_CHILD_ROLE1);
     I.click('Save');
     I.see('Success');
     I.waitForText('User details updated successfully');
     I.click('Return to user details');
     I.see('User Details');
-
-    const updatedRoles = await I.grabTextFromAll('[id^=\'assigned-role\']');
-    Assert.equal(updatedRoles.includes(INDEPENDANT_CHILD_ROLE), true);
-    Assert.equal(updatedRoles.includes(ASSIGNABLE_CHILD_ROLE), true);
+    I.see(INDEPENDANT_CHILD_ROLE);
+    I.see(ASSIGNABLE_CHILD_ROLE1);
   });
 
 Scenario('I as a user should be able to edit mfa if I have the assignable role of idam-mfa-disabled',
@@ -226,9 +223,7 @@ Scenario('I as a user should be able to edit mfa if I have the assignable role o
     I.fillField('#search', activeUserEmail);
     I.click('Search');
     I.waitForText('User Details');
-
-    let mfa = await I.grabTextFrom('#mfa');
-    Assert.equal(mfa.includes(MFA_DISABLED_TEXT), true);
+    I.see(MFA_DISABLED_TEXT);
 
     I.click('Edit user');
     I.waitForText('Edit User');
@@ -237,16 +232,14 @@ Scenario('I as a user should be able to edit mfa if I have the assignable role o
     const mfaFlag = await I.grabValueFrom(locate('//input[@name=\'multiFactorAuthentication\']'));
     Assert.equal(mfaFlag, MFA_ENABLED_FLAG);
 
-    I.click(MFA_ENABLED_TEXT);
+    I.click('Enabled');
     I.click('Save');
     I.see('Success');
     I.waitForText('User details updated successfully');
 
     I.click('Return to user details');
     I.see('User Details');
-
-    mfa = await I.grabTextFrom('#mfa');
-    Assert.equal(mfa.includes(MFA_ENABLED_TEXT), true);
+    I.see('ENABLED');
 
     I.click('Edit user');
     I.waitForText('Edit User');
@@ -268,9 +261,7 @@ Scenario('I as a user should not be able to edit mfa if I don\'t have the assign
     I.fillField('#search', PARENT_ROLE_EMAIL);
     I.click('Search');
     I.waitForText('User Details');
-
-    const mfa = await I.grabTextFrom('#mfa');
-    Assert.equal(mfa.includes(MFA_ENABLED_TEXT), true);
+    I.see('ENABLED');
 
     I.click('Edit user');
     I.waitForText('Edit User');
@@ -279,4 +270,41 @@ Scenario('I as a user should not be able to edit mfa if I don\'t have the assign
     // Check the mfa enabled checkbox is selected but disabled
     const mfaFlag = await I.grabValueFrom(locate('//input[@name=\'multiFactorAuthentication\' and @checked and @disabled]'));
     Assert.equal(mfaFlag, MFA_ENABLED_FLAG);
+  });
+
+Scenario('I as a user should be able to filter through roles while updating the user details',
+  {featureFlags: [BETA_EDIT]},
+  async ({I}) => {
+
+    const activeUserEmail = randomData.getRandomEmailAddress();
+    await I.createUserWithRoles(activeUserEmail, testConfig.PASSWORD, testConfig.USER_FIRSTNAME, [ASSIGNABLE_CHILD_ROLE1]);
+    const searchText = ASSIGNABLE_CHILD_ROLE2.substring(0, 10);
+
+    I.loginAs(PARENT_ROLE_EMAIL, testConfig.PASSWORD);
+    I.waitForText('Manage an existing user');
+    I.click('Manage an existing user');
+    I.click('Continue');
+    I.waitForText('Please enter the email address, user ID or SSO ID of the user you wish to manage');
+    I.click('#search');
+    I.fillField('#search', activeUserEmail);
+    I.click('Search');
+    I.waitForText('User Details');
+    I.click('Edit user');
+    I.waitForText('Edit User');
+    I.dontSee('Suspend user');
+    I.dontSee('Delete user');
+    I.see(ASSIGNABLE_CHILD_ROLE1);
+    I.see(ASSIGNABLE_CHILD_ROLE2);
+    I.see(PARENT_ROLE);
+    I.click('#roles__search-box');
+    I.fillField('#roles__search-box', searchText);
+
+    const checkboxes = await I.grabValueFromAll(locate('//div[@class=\'govuk-checkboxes__item\' and not(@hidden)]/input[@name=\'roles\']'));
+    checkboxes.forEach(function (checkbox) {
+      if (checkbox.includes(searchText) && !checkbox.includes(testConfig.RBAC.access)) {
+        Assert.ok(true);
+      } else {
+        Assert.ok(false);
+      }
+    });
   });
